requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.1.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.14.0
dbt-common==1.14.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
dlt==1.6.1
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
hexbytes==1.3.0
humanize==4.11.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.5
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
makefun==1.15.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
networkx==3.4.2
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.2.3
pendulum==3.0.0
pluggy==1.5.0
ply==3.11
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requirements-parser==0.11.0
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.22.3
semver==3.0.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.0.4
sqlparse==0.5.3
tenacity==9.0.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
tomlkit==0.13.2
types-requests==2.32.0.20241016
types-setuptools==75.8.0.20250110
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
